Added pizzas to menu(index page) and change database charset to UTF-8 and small changes in frontend index page

+++ src/main/java/pl/rasilewicz/restaurant_manager/services/ProductService.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface ProductService {
 
-    List<Product> findProductsByType(Long typeId);
+    List<Product> findProductsByType(Integer typeId);
 
     List<Product> findAllProducts();
 }
